Método Count agregado

+++ application/src/main/java/co/edu/javeriana/as/personapp/application/port/in/PersonInputPort.java
@@ -24,7 +24,7 @@ public interface PersonInputPort {
 
 	public Person findOne(Integer identification) throws NoExistException;
 
-	public Integer count();
+	public Long count();
 
 	public List<Phone> getPhones(Integer identification) throws NoExistException;
