+++ application/src/main/java/co/edu/javeriana/as/personapp/application/usecase/ProfessionUseCase.java
@@ -65,7 +65,7 @@ public class ProfessionUseCase implements ProfessionInputPort {
     }
 
     @Override
-    public Integer count() {
+    public long count() {
         return findAll().size();
     }
